- added mediaplayer service - added notification, locksceen music controller - improve ui

+++ app/src/main/java/player/project/com/musicplayer/Utilitys.java
@@ -73,7 +73,7 @@ public class Utilitys {
     }
 
 
-    public static int progressToTimer(int progress, int totalDuration) {
+    public static int progressToTimer(int progress, long totalDuration) {
         int currentDuration = 0;
         totalDuration = (int) (totalDuration / 1000);
         currentDuration = (int) ((((double) progress) / 100) * totalDuration);
